Ajout des cles primaires & MAJ Migrations

+++ database/migrations/2020_08_20_204609_create_tarifs_table.php
@@ -15,12 +15,13 @@ class CreateTarifsTable extends Migration
     {
         Schema::create('tarifs', function (Blueprint $table) {
             $table->CHAR('id_Tarif',5)->unique();
-            $table->NUMERIC('somme_Tarif');
+            $table->float('somme_Tarif');
             $table->string('annotations_Tarif');
             $table->timestamps();
 
             //Clé primaire
             $table->primary('id_Tarif');
+            
         });
     }
